+++ bin/devbox/vagrant/provision/scripts/connect_scripts/atlas-run-connect.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+/home/vagrant/scripts/oracle-license.sh
+
+node /home/vagrant/scripts/atlas-run-connect.js $@