+++ integration-tests/src/test/java/com/atlassian/plugin/connect/test/pageobjects/jira/JiraAdministrationHomePage.java
@@ -1,11 +1,7 @@
 package com.atlassian.plugin.connect.test.pageobjects.jira;
 
 import com.atlassian.jira.pageobjects.pages.AbstractJiraPage;
-import com.atlassian.pageobjects.elements.PageElement;
 import com.atlassian.pageobjects.elements.query.TimedCondition;
-import com.atlassian.plugin.connect.api.xmldescriptor.XmlDescriptor;
-import com.atlassian.plugin.connect.modules.util.ModuleKeyUtils;
-import com.atlassian.plugin.connect.test.pageobjects.ConnectAddOnEmbeddedTestPage;
 import org.openqa.selenium.By;
 
 public class JiraAdministrationHomePage extends AbstractJiraPage