+++ integration-tests/src/test/java/it/modules/confluence/TestDynamicContentMacro.java
@@ -1,9 +1,7 @@
 package it.modules.confluence;
 
 import com.atlassian.confluence.it.Page;
-import com.atlassian.confluence.pageobjects.component.dialog.MacroBrowserDialog;
-import com.atlassian.confluence.pageobjects.component.dialog.MacroForm;
-import com.atlassian.confluence.pageobjects.component.dialog.MacroItem;
+import com.atlassian.confluence.it.User;
 import com.atlassian.confluence.pageobjects.page.content.CreatePage;
 import com.atlassian.confluence.pageobjects.page.content.EditContentPage;
 import com.atlassian.plugin.connect.modules.beans.DynamicContentMacroModuleBean;
@@ -12,6 +10,7 @@ import com.atlassian.plugin.connect.test.AddonTestUtils;
 import com.atlassian.plugin.connect.test.pageobjects.RemotePluginDialog;
 import com.atlassian.plugin.connect.test.pageobjects.confluence.ConfluenceEditorContent;
 import com.atlassian.plugin.connect.test.pageobjects.confluence.ConfluenceOps;
+import com.atlassian.plugin.connect.test.pageobjects.confluence.ConfluencePageWithRemoteMacro;
 import com.atlassian.plugin.connect.test.pageobjects.confluence.RenderedMacro;
 import com.atlassian.plugin.connect.test.server.ConnectRunner;
 import it.servlet.ConnectAppServlets;
@@ -21,6 +20,14 @@ import org.apache.commons.lang.StringUtils;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import redstone.xmlrpc.XmlRpcFault;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.net.MalformedURLException;
 
 import static com.atlassian.fugue.Option.some;
 import static com.atlassian.plugin.connect.modules.beans.DynamicContentMacroModuleBean.newDynamicContentMacroModuleBean;
@@ -46,6 +53,9 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
 
     private static final String TABLE_MACRO_NAME = "Table Macro";
     private static final String TABLE_MACRO_KEY = "table-macro";
+
+    private static final String SLOW_MACRO_NAME = "Slow Macro";
+    private static final String SLOW_MACRO_KEY = "slow-macro";
 
     private static ConnectRunner remotePlugin;
 
@@ -105,6 +115,11 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
                 .withOutputType(MacroOutputType.BLOCK)
                 .build();
 
+        DynamicContentMacroModuleBean slowMacro = newDynamicContentMacroModuleBean()
+                .withUrl("/slow-macro")
+                .withKey(SLOW_MACRO_KEY)
+                .withName(new I18nProperty(SLOW_MACRO_NAME, null))
+                .build();
 
         remotePlugin = new ConnectRunner(product.getProductInstance().getBaseUrl(), AddonTestUtils.randomAddOnKey())
 
@@ -124,7 +139,8 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
                         hiddenMacro,
                         clientSideBodyEditingMacro,
                         clientSideBodyEditingMacroScriptInjection,
-                        macroInTableMacro
+                        macroInTableMacro,
+                        slowMacro
                 )
                 .addRoute(DEFAULT_MACRO_URL, ConnectAppServlets.helloWorldServlet())
                 .addRoute("/render-editor", ConnectAppServlets.macroEditor())
@@ -135,6 +151,7 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
                 .addRoute("/images/placeholder.png", ConnectAppServlets.resourceServlet("atlassian-icon-16.png", "image/png"))
                 .addRoute("/images/macro-icon.png", ConnectAppServlets.resourceServlet("atlassian-icon-16.png", "image/png"))
                 .addRoute("/render-macro-in-table-macro", ConnectAppServlets.apRequestServlet())
+                .addRoute("/slow-macro", new SlowMacroServlet(22))
                 .start();
     }
 
@@ -150,16 +167,13 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
     @Test
     public void testMacroIsRendered() throws Exception
     {
-        CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
-        editorPage.setTitle(randomName("Simple Macro on Page"));
+        testMacroIsRendered(TestUser.ADMIN.confUser());
+    }
 
-        selectMacro(editorPage, SIMPLE_MACRO_NAME);
-
-        savedPage = editorPage.save();
-        RenderedMacro renderedMacro = connectPageOperations.findMacroWithIdPrefix(SIMPLE_MACRO_KEY, 0);
-        String content = renderedMacro.getIFrameElementText("hello-world-message");
-
-        assertThat(content, is("Hello world"));
+    @Test
+    public void testMacroIsRenderedForAnonymous() throws Exception
+    {
+        testMacroIsRendered(null);
     }
 
     @Test
@@ -168,7 +182,7 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
         CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
         editorPage.setTitle(randomName("Short Body Macro"));
 
-        selectMacro(editorPage, SHORT_BODY_MACRO_NAME);
+        selectMacroAndSave(editorPage, SHORT_BODY_MACRO_NAME);
 
         ConfluenceEditorContent editorContent = (ConfluenceEditorContent) editorPage.getEditor().getContent();
         editorContent.setRichTextMacroBody("a short body");
@@ -186,7 +200,7 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
         CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
         editorPage.setTitle(randomName("Long Body Macro"));
 
-        selectMacro(editorPage, LONG_BODY_MACRO_NAME);
+        selectMacroAndSave(editorPage, LONG_BODY_MACRO_NAME);
 
         String body = StringUtils.repeat("x ", 200);
         ConfluenceEditorContent editorContent = (ConfluenceEditorContent) editorPage.getEditor().getContent();
@@ -205,13 +219,9 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
     {
         CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
         editorPage.setTitle(randomName("Parameter Page"));
-
-        MacroBrowserDialog macroBrowser = editorPage.openMacroBrowser();
-        MacroItem macro = macroBrowser.searchForFirst(PARAMETER_MACRO_NAME);
-        MacroForm macroForm = macro.select();
-
-        macroForm.getAutocompleteField("param1").setValue("param value");
-        macroBrowser.clickSave();
+        final MacroBrowserAndEditor macroBrowserAndEditor = selectMacro(editorPage, PARAMETER_MACRO_NAME);
+        macroBrowserAndEditor.macroForm.getAutocompleteField("param1").setValue("param value");
+        macroBrowserAndEditor.browserDialog.clickSave();
 
         savedPage = editorPage.save();
 
@@ -227,8 +237,8 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
         CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
         editorPage.setTitle(randomName("Multiple Macros"));
 
-        selectMacro(editorPage, SIMPLE_MACRO_NAME);
-        selectMacro(editorPage, SIMPLE_MACRO_NAME);
+        selectMacroAndSave(editorPage, SIMPLE_MACRO_NAME);
+        selectMacroAndSave(editorPage, SIMPLE_MACRO_NAME);
 
         savedPage = editorPage.save();
 
@@ -250,7 +260,7 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
         CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
         editorPage.setTitle(randomName("Small Inline Macro"));
 
-        selectMacro(editorPage, SMALL_INLINE_MACRO_NAME);
+        selectMacroAndSave(editorPage, SMALL_INLINE_MACRO_NAME);
 
         savedPage = editorPage.save();
 
@@ -264,13 +274,7 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
     {
         CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
         editorPage.setTitle(randomName("Macro Editor"));
-
-        MacroBrowserDialog macroBrowser = editorPage.openMacroBrowser();
-        MacroItem macro = macroBrowser.searchForFirst(EDITOR_MACRO_NAME);
-        macro.select();
-
-        RemotePluginDialog dialog = connectPageOperations.findDialog(EDITOR_MACRO_KEY);
-        dialog.submit();
+        selectMacro(editorPage, EDITOR_MACRO_NAME, macroDialogSubmitter(EDITOR_MACRO_KEY));
 
         savedPage = editorPage.save();
         RenderedMacro renderedMacro = connectPageOperations.findMacroWithIdPrefix(EDITOR_MACRO_KEY);
@@ -346,15 +350,93 @@ public class TestDynamicContentMacro extends AbstractContentMacroTest
         assertThat(content, is("body: <strong>must</strong> be removed:"));
     }
 
+    @Test
+    public void testSlowMacro() throws Exception
+    {
+        CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
+        editorPage.setTitle(randomName("Slow Macro on Page"));
+
+        selectMacroAndSave(editorPage, SLOW_MACRO_NAME);
+
+        savedPage = editorPage.save();
+        final ConfluencePageWithRemoteMacro page = product.visit(ConfluencePageWithRemoteMacro.class, savedPage.getTitle(), SLOW_MACRO_KEY);
+        assertThat(page.macroHasTimedOut(), is(true));
+    }
+
+    @Test
+    public void fastMacroShouldNotTimeOut() throws Exception
+    {
+        CreatePage editorPage = getProduct().loginAndCreatePage(TestUser.ADMIN.confUser(), TestSpace.DEMO);
+        editorPage.setTitle(randomName("Slow Macro on Page"));
+
+        selectMacroAndSave(editorPage, SIMPLE_MACRO_NAME);
+
+        savedPage = editorPage.save();
+        final ConfluencePageWithRemoteMacro page = product.visit(ConfluencePageWithRemoteMacro.class, savedPage.getTitle(), SLOW_MACRO_KEY);
+        assertThat(page.macroHasTimedOut(), is(false));
+    }
+
+    @Test
+    public void testMacroInComment() throws MalformedURLException, XmlRpcFault
+    {
+        addSimpleMacroToComment();
+
+        RenderedMacro renderedMacro = connectPageOperations.findMacroWithIdPrefix(SIMPLE_MACRO_KEY, 0);
+        String content = renderedMacro.getIFrameElementText("hello-world-message");
+
+        assertThat(content, is("Hello world"));
+    }
+
     @Override
     protected String getAddonBaseUrl()
     {
         return remotePlugin.getAddon().getBaseUrl();
     }
 
+    private void testMacroIsRendered(User user) throws Exception
+    {
+        getMacroContent(user, SIMPLE_MACRO_NAME, "Simple macro");
+        RenderedMacro renderedMacro = connectPageOperations.findMacroWithIdPrefix(SIMPLE_MACRO_KEY, 0);
+        String content = renderedMacro.getIFrameElementText("hello-world-message");
+
+        assertThat(content, is("Hello world"));
+    }
+
     private EditContentPage createAndEditPage(String pageName, String pageContent) throws Exception
     {
         ConfluenceOps.ConfluencePageData pageData = confluenceOps.setPage(some(TestUser.ADMIN), TestSpace.DEMO.getKey(), pageName, pageContent);
         return product.visit(EditContentPage.class, new Page(Long.parseLong(pageData.getId())));
+    }
+
+    private static final class SlowMacroServlet extends HttpServlet
+    {
+        private final int seconds;
+
+        private SlowMacroServlet(int seconds)
+        {
+            this.seconds = seconds;
+        }
+
+        @Override
+        protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
+        {
+            final long wakeTime = System.currentTimeMillis() + seconds * 1000 + 1;
+
+            while (System.currentTimeMillis() < wakeTime)
+            {
+                try
+                {
+                    Thread.sleep(wakeTime - System.currentTimeMillis());
+                }
+                catch (InterruptedException e)
+                {
+                    // do nothing
+                }
+            }
+
+            resp.setContentType("text/html");
+            resp.getWriter().write("finished");
+            resp.getWriter().close();
+        }
     }
 }